added passport authentication

+++ server/models/catModel.js
@@ -19,6 +19,7 @@
 // ];
 
 const pool = require("../database/db");
+const { addUser } = require("./userModel");
 const promisePool = pool.promise();
 
 const getAllCats = async (res) => {
@@ -56,15 +57,28 @@ const addCat = async (cat, res) => {
   }
 };
 
-const deleteCatById = async (catId, res) => {
+const deleteCatById = async (catId, owner, res) => {
+  if (userId == "4") {
   try {
-    const [rows] = await promisePool.query("DELETE FROM wop_cat WHERE cat_id = ?", [catId]);
+    const [rows] = await promisePool.query("DELETE FROM wop_cat WHERE cat_id = ? AND wop_cat.owner = wop_user.user_id = ?", [catId, owner]);
+    // cat_id = ? AND owner = ? AND EXISTS (SELECT * FROM wop_user WHERE wop_user.user_id = wop_cat.owner AND wop_user.role = 0)", [catId, owner]);
     return rows;
+
   } catch (e) {
     console.error("error", e.message);
     res.status(500).send(e.message);
   }
 };
+}
+// const deleteCatById = async (catId, owner, res) => {
+//   try {
+//     const [rows] = await promisePool.query("DELETE FROM wop_cat WHERE cat_id = ? AND owner = ?", [catId, owner]);
+//     return rows;
+//   } catch (e) {
+//     console.error("error", e.message);
+//     res.status(500).send(e.message);
+//   }
+// };
 
 const updateCatById = async (cat, res) => {
   try {
